requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
groq==0.11.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
